+++ test/core/CMakeLists.txt
@@ -24,3 +24,4 @@ add_subdirectory(subscription)
 add_subdirectory(telemetry)
 add_subdirectory(transaction_pool)
 add_subdirectory(dispute_coordinator)
+add_subdirectory(parachain)
