+++ test/core/crypto/crypto_store/CMakeLists.txt
@@ -16,7 +16,6 @@ addtest(crypto_store_test
 target_link_libraries(crypto_store_test
     crypto_store
     base_fs_test
-    validator_parachain
     )
 
 addtest(session_keys_test
